requirements.txt
faster-whisper
torch
sounddevice
scipy
langdetect
opencv-python
ultralytics
google-generativeai
gTTS
playsound==1.2.2
pyttsx3
